Initialize some proxies for Fault Proof on Sepolia
- `DisputeGameFactoryProxy`
- `DelayedWETHProxy`

`AnchorStateRegistryProxy` to follow in a separate set of changes
once `DisputeGameFactoryProxy` is ready as the `AnchorStateRegistry`
implementation depends on it.

These are prerequisites for Fault Proof for Base Sepolia.

The factory will not be useable just yet, until we `setImplementation`
for specific game types. The game implementations will be deployed
separately as they depend on the anchor state registry.

The proxies were deployed using our internal Smart Contract Manager (SCM).
We will need the proxy admin to initialize these currently blank proxies.